I changed a thing about the .ui file and now it works. I changed a thing about the .ui file *back*, and it still works.
I hate this.
Suspect that the schema.xml is supposed to be compiled and put in place by "gnome-extensions install", but this is A LIE. Better off just shipping your own .compiled file. Why does gnome even have this? There are plenty of other places where you don't mind ingesting a text xml file. Why does this doodad have to be so special?